# src/authserver/web/openapi.yaml
openapi: 3.0.3
info:
  title: Goiabada API
  description: |
    REST API for Goiabada authentication and authorization server.

    ## Authentication

    Every operation under `/api/v1` requires a valid JWT access token in the
    `Authorization` header:
    ```
    Authorization: Bearer <access_token>
    ```
    `getClientLogoImage` is the exception and is marked `security: []`: it is public.

    ## API Groups

    - **Admin API** (`/api/v1/admin/*`): administrative control. Each operation accepts any
      one of a small set of scopes rather than a single one: reads accept
      `authserver:admin-read` alongside the domain's own scope, and writes accept the
      domain scope (`authserver:manage-users`, `authserver:manage-clients` or
      `authserver:manage-settings`). `authserver:manage` is accepted everywhere.
      A token from the client credentials flow is accepted here.
    - **Account API** (`/api/v1/account/*`): Self-service account management. Requires
      `authserver:manage-account` scope, **and** a token issued for a user: these operations
      resolve the acting user from `sub`, so a client credentials token is refused with 403
      and `error_code` `USER_CONTEXT_REQUIRED`.

    ## Caching

    Every response from an operation in either API group carries these headers, error
    responses and the token and scope refusals included:
    ```
    Cache-Control: no-store
    Pragma: no-cache
    ```

...

...
    none of them is described by this document.

...
servers:
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/admin/users/{id}/profile-picture:
    get:
      tags: [Users]
      summary: Get user profile picture info
      description: Check whether a user has a profile picture and get the picture URL
      operationId: getUserProfilePictureInfo
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
      responses:
        '200':
          description: Profile picture info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfilePictureInfoResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags: [Users]
      summary: Upload user profile picture
      description: Upload or replace the user's profile picture. Supports JPEG, PNG, GIF, and WebP formats up to the configured maximum size.
      operationId: uploadUserProfilePicture
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - picture
              properties:
                picture:
                  type: string
                  format: binary
                  description: Profile picture file (JPEG, PNG, GIF, or WebP)
      responses:
        '200':
          description: Profile picture uploaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfilePictureUploadResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags: [Users]
      summary: Delete user profile picture
      description: Remove the user's profile picture
      operationId: deleteUserProfilePicture
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
      responses:
        '200':
          description: Profile picture deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/v1/admin/groups:
    get:
      tags: [Groups]
      summary: List groups
      description: Get all groups
      operationId: getGroups
      responses:
        '200':
          description: Groups list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetGroupsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags: [Groups]
      summary: Create group
      description: Create a new group
      operationId: createGroup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGroupRequest'
      responses:
        '201':
          description: Group created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateGroupResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/v1/admin/groups/{id}:
    get:
      tags: [Groups]
      summary: Get group
      description: Get group details by ID
      operationId: getGroup
      parameters:
        - $ref: '#/components/parameters/GroupIdParam'
      responses:
        '200':
          description: Group details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetGroupResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags: [Groups]
      summary: Update group
      description: Update group details
      operationId: updateGroup
      parameters:
        - $ref: '#/components/parameters/GroupIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateGroupRequest'
      responses:
        '200':
          description: Group updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateGroupResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags: [Groups]
      summary: Delete group
      description: Delete a group
      operationId: deleteGroup
      parameters:
        - $ref: '#/components/parameters/GroupIdParam'
      responses:
        '200':
          description: Group deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token.

        Admin API: any one of `authserver:admin-read`, the domain scope for the operation
        (`authserver:manage-users`, `authserver:manage-clients`, `authserver:manage-settings`)
        or `authserver:manage`. A client credentials token is accepted.

        Account API: scope `authserver:manage-account`, on a token issued for a user. A client
        credentials token carries no `auth_time` claim and is refused with 403
        `USER_CONTEXT_REQUIRED`.

        Browser sessions: scope `authserver:browser-sessions`, and only that scope. It is
        deliberately not one of the manage-* scopes, so holding the admin console's client secret
        is not a way to drive the admin API with no user present.
  parameters:
    UserIdParam:
      name: id
      in: path
      required: true
      description: User ID
      schema:
        type: integer
        format: int64
    GroupIdParam:
      name: id
      in: path
      required: true
      description: Group ID
      schema:
        type: integer
        format: int64
...
  responses:
    Unauthorized:
      description: Missing or invalid bearer token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    BadRequest:
      description: Invalid request body or parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: |
        The token is valid but does not authorise this request. Three checks
        answer here, and a caller distinguishes them by error_code.

        INSUFFICIENT_SCOPE comes from the route's authorisation middleware,
        before the handler runs: an Admin operation needs one of the scopes
        its route names, and an Account operation needs
        `authserver:manage-account`.

        USER_CONTEXT_REQUIRED comes from the same middleware on the Account
        operations, which also require the token to have been issued for a
        user: one obtained through the client credentials grant carries no
        auth_time claim and is refused here.

        FORBIDDEN comes from the handler, on the account endpoints that
        resolve a resource by id and then check it belongs to the caller, so
        a valid token for the wrong user reaches this rather than a 404.

        Every operation this document secures declares a 403. The public
        client logo endpoint, which carries `security: []` and authorises
        nothing, is the one that does not.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalServerError:
      description: |
        The request failed inside the server, most often a storage error.
        error_code is INTERNAL_SERVER_ERROR, except on key rotation, which also
        answers KEY_SET_INCOMPLETE. The description carries no detail about the
        failure, only the request id an operator can find the log line by.
        Retrying is reasonable.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ErrorResponse:
      type: object
      description: |
        Admin/account API error envelope. Routing by HTTP status code:
        4xx is a client-correctable failure (validation, not found,
        permission denied, rate limited); 5xx is a server-side problem.
        The body provides a stable error_code, optional error_args, and
        an English error_description for non-localizing consumers.
        Protocol endpoints (/auth/token, /auth/authorize, /connect/register,
        /userinfo) use their RFC-defined error envelopes instead of this one.
      properties:
        error_code:
          type: string
          description: |
            Stable identifier for the failure. UPPER_SNAKE for legacy codes
            (e.g. VALIDATION_ERROR), dotted lowercase for catalog-keyed
            localized codes (e.g. validator.email.required).
        error_args:
          type: object
          additionalProperties: true
          description: Substitution values for the localized message template.
        error_description:
          type: string
          description: Rendered English message. Stable for non-localizing consumers.
      required:
        - error_description
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success

    # =========================================================================
    # User Schemas
    # =========================================================================
...
    GroupResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
          nullable: true
        updatedAt:
          type: string
          format: date-time
          nullable: true
        groupIdentifier:
          type: string
        description:
          type: string
        includeInIdToken:
          type: boolean
        includeInAccessToken:
          type: boolean
        memberCount:
          type: integer
      required:
        - id
        - createdAt
        - updatedAt
        - groupIdentifier
        - description
        - includeInIdToken
        - includeInAccessToken
        - memberCount
    GetGroupsResponse:
      type: object
      properties:
        groups:
          type: array
          items:
            $ref: '#/components/schemas/GroupResponse'
      required:
        - groups
    GetGroupResponse:
      type: object
      properties:
        group:
          $ref: '#/components/schemas/GroupResponse'
      required:
        - group
    CreateGroupRequest:
      type: object
      required:
        - groupIdentifier
      properties:
        groupIdentifier:
          type: string
        description:
          type: string
          maxLength: 100
        includeInIdToken:
          type: boolean
        includeInAccessToken:
          type: boolean
    CreateGroupResponse:
      type: object
      properties:
        group:
          $ref: '#/components/schemas/GroupResponse'
      required:
        - group
    UpdateGroupRequest:
      type: object
      properties:
        groupIdentifier:
          type: string
        description:
          type: string
          maxLength: 100
        includeInIdToken:
          type: boolean
        includeInAccessToken:
          type: boolean
    UpdateGroupResponse:
      type: object
      properties:
        group:
          $ref: '#/components/schemas/GroupResponse'
      required:
        - group
...
    ProfilePictureInfoResponse:
      type: object
      properties:
        hasPicture:
          type: boolean
        pictureUrl:
          type: string
          description: Public URL to serve the picture (only present when hasPicture is true)
      required:
        # pictureUrl is deliberately absent: the handler adds that key only when hasPicture is
        # true.
        - hasPicture
    ProfilePictureUploadResponse:
      type: object
      properties:
        success:
          type: boolean
        pictureUrl:
          type: string
          description: Public URL to the uploaded picture
      required:
        - success
        - pictureUrl

    # =========================================================================
    # Settings Schemas
    # =========================================================================
...